Move new page to index
